requirements.txt
setuptools
requests
python-dotenv
termcolor
pandas
click